Home.vueから一部スクリプトを分離独立
