+++ package.json
@@ -11,8 +11,10 @@
   },
   "dependencies": {
     "core-js": "^3.11.2",
+    "dompurify": "^3.3.3",
     "vue": "^3.5.24",
-    "vue-router": "^4.6.4"
+    "vue-router": "^4.6.4",
+    "vue-tsc": "^3.2.5"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.28.6",
